[Chore] Group the export pipeline's tests by what they exercise
One 2 903-line test module covering everything the pipeline does. Now
five files, chosen by which production functions each test actually
calls rather than by name:

    naming    output paths, export groups, type selectors    19 tests
    reading   reading objects out of a bundle, and limits     22
    payload   writing a payload, and telling duplicates apart 31
    images    decoding and encoding images                     6
    media     post-processing audio and video                  8

Fixtures move to `support` rather than being copied. That matters more
here than in the other splits: several of them build configs and payload
bytes the assertions compare against, so two drifting copies would not
fail -- they would quietly test different things.

Same 202 tests before and after, and the same count of individual `ok`
lines, so nothing was dropped or duplicated in the move.
